--- backend/myMath/max.go
@@ -1,9 +1,0 @@
-package myMath
-
-func MaxInt(x, y int) int {
-	if x >= y {
-		return x
-	} else {
-		return y
-	}
-}
