--- backend/submission/submissionQueue.go
@@ -1,17 +1,0 @@
-package submission
-
-import (
-	"go-test/user"
-
-	"github.com/gin-gonic/gin"
-)
-
-func SubmissionQueueHandler(c *gin.Context) {
-	userName := c.Param("userName")
-	userData := user.GetUserFromUserName(c, userName)
-	submissions, err := GetSubmissionsFromUser(c, *userData)
-	if err != nil {
-		c.JSON(400, err.Error())
-	}
-	c.JSON(200, submissions)
-}
