+++ backend/util/max.go
@@ -1,0 +1,9 @@
+package util
+
+func MaxInt(x, y int) int {
+	if x >= y {
+		return x
+	} else {
+		return y
+	}
+}
