+++ frontend/src/App.tsx
@@ -3,13 +3,13 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { AxiosClientProvider } from "./providers/AxiosClientProvider";
 
 import Login from "./routes/Login";
-import Account from "./routes/Account";
-import Dashboard from "./routes/Dashboard";
-import Message from "./routes/Message";
-import Schedule from "./routes/Schedule";
-import AssignmentPage from "./routes/AssignmentPage";
-import ResultQueuePage from "./routes/ResultQueuePage";
-import SubmitDetailsPage from "./routes/SubmitDetailsPage";
+import AccountPage from "./routes/Account";
+import DashboardPage from "./routes/Dashboard";
+import MessagePage from "./routes/Message";
+import SchedulePage from "./routes/Schedule";
+import ProblemPage from "./routes/Problem";
+import ResultsPage from "./routes/Results";
+import SubmissionPage from "./routes/Submission";
 import NotFound from "./routes/NotFound";
 import { RouteAuthGuard } from "./providers/RouteAuthGuard";
 import { PageType } from "./types/PageTypes";
@@ -21,13 +21,13 @@ const App: React.FC = () => {
         <Routes>
           <Route path="/login" element={<Login />} />
           <Route path="/:userName">
-            <Route path="account" element={<RouteAuthGuard component={<Account />} pageType={PageType.Private} />} />
-            <Route path="dashboard" element={<RouteAuthGuard component={<Dashboard />} pageType={PageType.Private} />} />
-            <Route path="message" element={<RouteAuthGuard component={<Message />} pageType={PageType.Private} />} />
-            <Route path="schedule" element={<RouteAuthGuard component={<Schedule />} pageType={PageType.Private} />} />
-            <Route path="assignmentInfo/:problemId" element={<RouteAuthGuard component={<AssignmentPage />} pageType={PageType.Private} />} />
-            <Route path="results" element={<RouteAuthGuard component={<ResultQueuePage />} pageType={PageType.Private} />} />
-            <Route path="submission/:submissionId" element={<RouteAuthGuard component={<SubmitDetailsPage />} pageType={PageType.Private} />} />
+            <Route path="account" element={<RouteAuthGuard component={<AccountPage />} pageType={PageType.Private} />} />
+            <Route path="dashboard" element={<RouteAuthGuard component={<DashboardPage />} pageType={PageType.Private} />} />
+            <Route path="message" element={<RouteAuthGuard component={<MessagePage />} pageType={PageType.Private} />} />
+            <Route path="schedule" element={<RouteAuthGuard component={<SchedulePage />} pageType={PageType.Private} />} />
+            <Route path="problem/:problemId" element={<RouteAuthGuard component={<ProblemPage />} pageType={PageType.Private} />} />
+            <Route path="results" element={<RouteAuthGuard component={<ResultsPage />} pageType={PageType.Private} />} />
+            <Route path="submission/:submissionId" element={<RouteAuthGuard component={<SubmissionPage />} pageType={PageType.Private} />} />
           </Route>
           <Route path="*" element={<NotFound />} />
         </Routes>
